2lab/requirements.txt
alembic==1.14.1
fastapi==0.115.12
pydantic==2.10.6
PyJWT==2.9.0
python-multipart==0.0.20
SQLAlchemy==2.0.39
uvicorn==0.33.0
